pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "slacraper"
description = "Slack message scraper tool"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Communications :: Chat",
    "Topic :: Utilities",
]
requires-python = ">=3.12.0"
dependencies = [
    "slack-sdk>=3.0.0",
    "click>=8.0.0",
    "python-dateutil>=2.8.2",
    "pytest>=7.0.0",
]
dynamic = ["version"]

[tool.setuptools.package-dir]
slacraper = "src/slacraper"

[tool.setuptools_scm]
write_to = "src/slacraper/_version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
